+++ test/test_helper.exs
@@ -73,6 +73,8 @@ ExUnit.start(
     # :issuing_withdrawal_log,
     # :webhook,
     # :brcode_preview
+    # :individual_identity,
+    # :individual_identity_log,
   ]
 )
 
@@ -94,3 +96,4 @@ Code.require_file("./test/utils/ledger_transaction.exs")
 Code.require_file("./test/utils/issuing_holder.exs")
 Code.require_file("./test/utils/issuing_withdrawal.exs")
 Code.require_file("./test/utils/issuing_invoice.exs")
+Code.require_file("./test/utils/individual_identity.exs")